+++ src/lawEnforcement/LoadBgb.jsx
@@ -12,10 +12,11 @@ const LoadBgb = () => {
     const [bgbs, setBgbs] = useState(loadedBgbInfo);
 
     const handleDelete = (_id) => {
-        console.log(_id);
-        fetch(`https://polling-station-management-server.vercel.app/bgb/${_id}`, {
+         const confirmDelete = window.confirm("আপনি কি ডেটাটি মুছতে চান?");
+          if (confirmDelete) {
+            fetch(`https://polling-station-management-server.vercel.app/bgb/${_id}`, {
             method: "DELETE",
-        })
+            })
             .then((res) => res.json())
             .then((data) => {
                 // console.log(data);
@@ -27,6 +28,7 @@ const LoadBgb = () => {
                     setBgbs(remainingData);
                 }
             });
+          }
     };
 
     return (
